# Streaming PromQL engine: native histograms