--- a/pkg/streamingpromql/benchmarks/ingester.go
+++ b/pkg/streamingpromql/benchmarks/ingester.go
@@ -211,6 +211,9 @@ func pushTestData(ing *ingester.Ingester, metricSizes []int) error {
 	ctx := user.InjectOrgID(context.Background(), UserID)
 
 	// Batch samples into separate requests
+	// There is no precise science behind this number currently.
+	// A quick run locally found batching by 100 did not increase the loading time by any noticable amount.
+	// Additionally memory usage maxed about 4GB for the whole process.
 	batchSize := 100
 	for i := 0; i < NumIntervals; i += batchSize {
 		end := i + batchSize