--- a/pkg/streamingpromql/operator/instant_vector_selector.go
+++ b/pkg/streamingpromql/operator/instant_vector_selector.go
@@ -51,7 +51,7 @@ func (v *InstantVectorSelector) NextSeries(_ context.Context) (InstantVectorSeri
 
 	for stepT := v.Selector.Start; stepT <= v.Selector.End; stepT += v.Selector.Interval {
 		var t int64
-		var val float64
+		var f float64
 		var h *histogram.FloatHistogram
 
 		ts := stepT
@@ -67,7 +67,7 @@ func (v *InstantVectorSelector) NextSeries(_ context.Context) (InstantVectorSeri
 				return InstantVectorSeriesData{}, v.memoizedIterator.Err()
 			}
 		case chunkenc.ValFloat:
-			t, val = v.memoizedIterator.At()
+			t, f = v.memoizedIterator.At()
 		case chunkenc.ValHistogram:
 			t, h = v.memoizedIterator.AtFloatHistogram()
 		default:
@@ -76,31 +76,31 @@ func (v *InstantVectorSelector) NextSeries(_ context.Context) (InstantVectorSeri
 
 		if valueType == chunkenc.ValNone || t > ts {
 			var ok bool
-			t, val, h, ok = v.memoizedIterator.PeekPrev()
+			t, f, h, ok = v.memoizedIterator.PeekPrev()
 			if !ok || t < ts-v.Selector.LookbackDelta.Milliseconds() {
 				continue
 			}
 		}
-		if value.IsStaleNaN(val) || (h != nil && value.IsStaleNaN(h.Sum)) {
+		if value.IsStaleNaN(f) || (h != nil && value.IsStaleNaN(h.Sum)) {
 			continue
 		}
 
-		// if (val, h have been set by PeekPrev, we do not know if val is 0 because that's the actual value, or because
+		// if (f, h) have been set by PeekPrev, we do not know if f is 0 because that's the actual value, or because
 		// the previous value had a histogram.
 		// PeekPrev will set the histogram to nil, or the value to 0 if the other type exists.
 		// So check if histograms is nil first. If we don't have a histogram, then we should have a value and vice-versa.
 		if h != nil {
 			if len(data.Histograms) == 0 {
-				// Only create the pool once we know the series is a histogram or not
+				// Only create the slice once we know the series is a histogram or not
 				data.Histograms = GetHPointSlice(v.numSteps)
 			}
 			data.Histograms = append(data.Histograms, promql.HPoint{T: stepT, H: h})
 		} else {
 			if len(data.Floats) == 0 {
-				// Only create the pool once we know the series is a histogram or not
+				// Only create the slice once we know the series is a histogram or not
 				data.Floats = GetFPointSlice(v.numSteps)
 			}
-			data.Floats = append(data.Floats, promql.FPoint{T: stepT, F: val})
+			data.Floats = append(data.Floats, promql.FPoint{T: stepT, F: f})
 		}
 	}