--- a/pkg/streamingpromql/testdata/ours/native_histograms.test
+++ b/pkg/streamingpromql/testdata/ours/native_histograms.test
@@ -3,13 +3,6 @@
 # Provenance-includes-license: Apache-2.0
 # Provenance-includes-copyright: The Prometheus Authors
 
-# Minimal valid case: an empty histogram.
-load 5m
-	empty_histogram	{{}}
-
-eval instant at 5m empty_histogram
-	{__name__="empty_histogram"} {{}}
-
 # buckets:[1 2 1] means 1 observation in the 1st bucket, 2 observations in the 2nd and 1 observation in the 3rd (total 4).
 load 5m
 	single_histogram	{{schema:0 sum:5 count:4 buckets:[1 2 1]}}