--- a/pkg/streamingpromql/testdata/upstream/native_histograms.test
+++ b/pkg/streamingpromql/testdata/upstream/native_histograms.test
@@ -144,9 +144,8 @@ load 5m
 # 	{} 1.5
 
 
-# Unsupported by streaming engine.
-# eval instant at 50m incr_histogram
-# 	{__name__="incr_histogram"} {{count:14 sum:24 buckets:[1 12 1]}}
+eval instant at 50m incr_histogram
+	{__name__="incr_histogram"} {{count:14 sum:24 buckets:[1 12 1]}}
 
 # Unsupported by streaming engine.
 # eval instant at 50m histogram_count(incr_histogram)
@@ -189,9 +188,8 @@ load 5m
 load 5m
 	low_res_histogram	{{schema:-1 sum:4 count:1 buckets:[1] offset:1}}+{{schema:0 sum:4 count:4 buckets:[2 2] offset:1}}x1
 
-# Unsupported by streaming engine.
-# eval instant at 5m low_res_histogram
-# 	{__name__="low_res_histogram"} {{schema:-1 count:5 sum:8 offset:1 buckets:[5]}}
+eval instant at 5m low_res_histogram
+	{__name__="low_res_histogram"} {{schema:-1 count:5 sum:8 offset:1 buckets:[5]}}
 
 # Unsupported by streaming engine.
 # eval instant at 5m histogram_count(low_res_histogram)