@jhesketh 2024-05-14
I've marked this as ready-for-review if we want to put this in in its current state. I'm also happy to keep adding commits to this branch as we support more histogram operations.