@jhesketh 2024-05-28
Benchmark of this branch in details:

<details>
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
                                                                       │      Prometheus      │              streaming              │
                                                                       │        sec/op        │    sec/op     vs base               │
Query/a_1,_instant_query                                                         148.0µ ±  4%   133.4µ ±  6%   -9.88% (p=0.002 n=6)
Query/a_1,_range_query_with_100_steps                                            158.1µ ±  3%   146.0µ ±  1%   -7.66% (p=0.002 n=6)
Query/a_1,_range_query_with_1000_steps                                           260.2µ ±  1%   242.8µ ±  4%   -6.66% (p=0.002 n=6)
Query/a_100,_instant_query                                                       1.071m ±  2%   1.047m ±  1%   -2.22% (p=0.015 n=6)
Query/a_100,_range_query_with_100_steps                                          2.301m ± 22%   1.798m ±  2%  -21.86% (p=0.009 n=6)
Query/a_100,_range_query_with_1000_steps                                         9.551m ±  1%   9.691m ±  3%   +1.47% (p=0.026 n=6)
Query/a_2000,_instant_query                                                      18.54m ±  1%   18.31m ±  3%        ~ (p=0.132 n=6)
Query/a_2000,_range_query_with_100_steps                                         32.44m ±  1%   32.01m ±  1%        ~ (p=0.065 n=6)
Query/a_2000,_range_query_with_1000_steps                                        156.2m ±  2%   155.4m ±  2%   -0.51% (p=0.041 n=6)
Query/nh_1,_instant_query                                                        246.6µ ±  7%   226.4µ ±  4%   -8.16% (p=0.002 n=6)
Query/nh_1,_range_query_with_100_steps                                           328.0µ ±  2%   319.7µ ±  5%        ~ (p=0.240 n=6)
Query/nh_1,_range_query_with_1000_steps                                          775.3µ ±  1%   953.6µ ± 20%        ~ (p=0.065 n=6)
Query/nh_100,_instant_query                                                      2.404m ±  2%   2.369m ±  2%        ~ (p=0.065 n=6)
Query/nh_100,_range_query_with_100_steps                                         9.274m ± 19%   7.308m ±  2%  -21.19% (p=0.002 n=6)
Query/nh_100,_range_query_with_1000_steps                                        57.76m ±  3%   58.53m ±  4%        ~ (p=0.589 n=6)
Query/nh_2000,_instant_query                                                     46.69m ±  2%   46.17m ±  3%        ~ (p=0.310 n=6)
Query/nh_2000,_range_query_with_100_steps                                        146.2m ±  3%   147.8m ±  2%        ~ (p=0.818 n=6)
Query/nh_2000,_range_query_with_1000_steps                                        1.309 ±  2%    1.289 ±  2%   -1.56% (p=0.026 n=6)
Query/a_1[1m],_instant_query                                                     203.8µ ±  3%   180.5µ ±  6%  -11.41% (p=0.002 n=6)
Query/a_100[1m],_instant_query                                                   583.6µ ±  2%   572.6µ ±  1%   -1.89% (p=0.002 n=6)
Query/a_2000[1m],_instant_query                                                  6.755m ±  3%   5.729m ±  1%  -15.19% (p=0.002 n=6)
Query/rate(a_1[1m]),_instant_query                                               205.1µ ±  3%   184.2µ ±  4%  -10.16% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                                  233.1µ ±  5%   200.7µ ±  5%  -13.91% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                                 426.0µ ±  3%   338.0µ ±  3%  -20.66% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                             765.8µ ±  2%   579.7µ ±  3%  -24.30% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                                2.519m ±  3%   1.764m ±  2%  -29.97% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                               16.34m ±  3%   10.33m ±  2%  -36.77% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                            9.358m ±  2%   6.639m ±  1%  -29.05% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                               34.88m ± 13%   25.84m ±  2%  -25.92% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_1000_steps                              253.8m ±  1%   157.4m ±  3%  -37.99% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_10000_steps                                1.963m ±  2%   1.421m ±  1%  -27.60% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_10000_steps                             128.81m ±  2%   81.30m ±  1%  -36.89% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_10000_steps                              2.726 ±  1%    1.630 ±  2%  -40.21% (p=0.002 n=6)
Query/rate(a_1[1d]),_instant_query                                               1.441m ±  3%   1.099m ±  2%  -23.75% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_100_steps                                  2.120m ±  2%   1.598m ±  2%  -24.65% (p=0.002 n=6)
Query/rate(a_1[1d]),_range_query_with_1000_steps                                 7.155m ±  2%   5.781m ±  2%  -19.20% (p=0.002 n=6)
Query/rate(a_100[1d]),_instant_query                                             64.16m ±  2%   54.78m ±  2%  -14.62% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_100_steps                                118.5m ±  1%   100.0m ±  2%  -15.61% (p=0.002 n=6)
Query/rate(a_100[1d]),_range_query_with_1000_steps                               577.0m ±  2%   504.2m ±  2%  -12.61% (p=0.002 n=6)
Query/rate(a_2000[1d]),_instant_query                                           1133.0m ±  3%   959.7m ±  1%  -15.30% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_100_steps                                2.197 ±  1%    1.853 ±  2%  -15.67% (p=0.002 n=6)
Query/rate(a_2000[1d]),_range_query_with_1000_steps                              11.433 ±  1%    9.840 ±  1%  -13.93% (p=0.002 n=6)
Query/a_1_-_b_1,_instant_query                                                   357.2µ ±  3%   326.6µ ±  6%   -8.55% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps                                      407.2µ ±  1%   349.0µ ±  2%  -14.29% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps                                     891.0µ ±  3%   559.4µ ±  2%  -37.21% (p=0.002 n=6)
Query/a_100_-_b_100,_instant_query                                               2.338m ±  1%   2.323m ±  2%        ~ (p=0.180 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps                                  6.159m ±  2%   3.826m ±  2%  -37.89% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps                                 36.81m ±  1%   16.17m ±  4%  -56.09% (p=0.002 n=6)
Query/a_2000_-_b_2000,_instant_query                                             33.00m ±  2%   32.07m ±  2%   -2.81% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps                               106.58m ±  1%   56.55m ±  1%  -46.94% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps                               818.0m ±  1%   276.8m ±  2%  -66.16% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps                                    4.875m ±  2%   2.415m ±  1%  -50.46% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps                                338.5m ±  2%   134.8m ±  1%  -60.17% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps                               8.693 ±  0%    2.675 ±  1%  -69.23% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query                                2.219m ±  2%   2.204m ±  3%        ~ (p=0.394 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps                   4.818m ±  3%   3.377m ±  1%  -29.90% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps                  26.86m ±  3%   14.42m ±  2%  -46.32% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query                                  25.30m ±  2%   21.53m ±  1%  -14.90% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps                     72.59m ±  4%   36.64m ±  1%  -49.53% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps                    517.3m ±  2%   173.2m ±  2%  -66.52% (p=0.002 n=6)
Query/sum(a_1),_instant_query                                                    181.9µ ±  2%   162.6µ ±  1%  -10.63% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_100_steps                                       202.3µ ±  1%   177.8µ ±  2%  -12.10% (p=0.002 n=6)
Query/sum(a_1),_range_query_with_1000_steps                                      345.1µ ±  3%   295.1µ ±  2%  -14.50% (p=0.002 n=6)
Query/sum(a_100),_instant_query                                                  1.193m ±  2%   1.158m ±  1%   -2.98% (p=0.004 n=6)
Query/sum(a_100),_range_query_with_100_steps                                     1.989m ±  2%   1.925m ±  2%   -3.18% (p=0.004 n=6)
Query/sum(a_100),_range_query_with_1000_steps                                    8.729m ±  5%   8.543m ±  2%   -2.14% (p=0.009 n=6)
Query/sum(a_2000),_instant_query                                                 16.12m ±  3%   15.73m ±  3%   -2.42% (p=0.041 n=6)
Query/sum(a_2000),_range_query_with_100_steps                                    29.80m ±  2%   28.41m ±  2%   -4.68% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_1000_steps                                   153.5m ±  2%   143.7m ±  3%   -6.41% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_instant_query                                             249.1µ ±  4%   236.5µ ±  4%   -5.07% (p=0.009 n=6)
Query/sum_by_(l)(h_1),_range_query_with_100_steps                                316.3µ ±  2%   292.5µ ±  3%   -7.51% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_range_query_with_1000_steps                               874.3µ ±  2%   813.3µ ±  2%   -6.98% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_instant_query                                           5.650m ±  1%   5.559m ±  1%   -1.62% (p=0.015 n=6)
Query/sum_by_(l)(h_100),_range_query_with_100_steps                             10.083m ±  2%   9.740m ±  1%   -3.40% (p=0.002 n=6)
Query/sum_by_(l)(h_100),_range_query_with_1000_steps                             51.21m ±  2%   46.22m ±  2%   -9.75% (p=0.002 n=6)
Query/sum_by_(l)(h_2000),_instant_query                                          96.32m ±  2%   95.69m ±  3%        ~ (p=0.180 n=6)
Query/sum_by_(l)(h_2000),_range_query_with_100_steps                             177.6m ±  2%   169.4m ±  1%   -4.63% (p=0.002 n=6)
Query/sum_by_(l)(h_2000),_range_query_with_1000_steps                           1154.6m ±  2%   870.4m ±  1%  -24.61% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_instant_query                                            250.0µ ±  3%   231.1µ ±  2%   -7.55% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_range_query_with_100_steps                               319.4µ ±  3%   301.1µ ±  1%   -5.73% (p=0.002 n=6)
Query/sum_by_(le)(h_1),_range_query_with_1000_steps                              900.7µ ±  1%   840.5µ ±  1%   -6.67% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_instant_query                                          5.448m ±  1%   5.335m ±  2%   -2.06% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_range_query_with_100_steps                             9.671m ±  1%   9.334m ±  2%   -3.48% (p=0.002 n=6)
Query/sum_by_(le)(h_100),_range_query_with_1000_steps                            48.98m ±  2%   46.28m ±  1%   -5.51% (p=0.002 n=6)
Query/sum_by_(le)(h_2000),_instant_query                                         95.01m ±  1%   91.96m ±  1%   -3.21% (p=0.002 n=6)
Query/sum_by_(le)(h_2000),_range_query_with_100_steps                            177.5m ±  2%   166.3m ±  2%   -6.28% (p=0.002 n=6)
Query/sum_by_(le)(h_2000),_range_query_with_1000_steps                          1106.8m ±  1%   857.5m ±  4%  -22.52% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query                               355.9µ ±  2%   316.2µ ±  3%  -11.16% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps                  433.8µ ±  1%   349.7µ ±  2%  -19.39% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps                1002.2µ ±  2%   600.3µ ±  1%  -40.11% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query                           1.460m ±  2%   1.134m ±  1%  -22.31% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps              6.636m ±  1%   3.205m ±  2%  -51.71% (p=0.002 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps             48.90m ±  1%   18.59m ±  1%  -61.99% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query                         17.70m ±  1%   13.02m ±  3%  -26.45% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps           118.20m ±  2%   45.63m ±  5%  -61.40% (p=0.002 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps          1083.7m ±  2%   317.8m ±  2%  -70.67% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_instant_query                                              366.7µ ±  2%   336.7µ ±  3%   -8.19% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_range_query_with_100_steps                                 422.4µ ±  2%   356.8µ ±  2%  -15.52% (p=0.002 n=6)
Query/sum(a_1_+_b_1),_range_query_with_1000_steps                                945.3µ ±  2%   581.9µ ±  2%  -38.44% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_instant_query                                          2.347m ±  1%   2.296m ±  1%   -2.16% (p=0.009 n=6)
Query/sum(a_100_+_b_100),_range_query_with_100_steps                             6.101m ±  1%   3.793m ±  3%  -37.83% (p=0.002 n=6)
Query/sum(a_100_+_b_100),_range_query_with_1000_steps                            37.57m ±  2%   16.63m ±  2%  -55.74% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_instant_query                                        33.04m ±  2%   31.80m ±  2%   -3.76% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_100_steps                          107.76m ±  1%   57.14m ±  1%  -46.97% (p=0.002 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_1000_steps                          840.0m ±  1%   281.6m ±  1%  -66.48% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_instant_query                                  224.8µ ±  2%   191.8µ ±  6%  -14.70% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_100_steps                     358.3µ ±  2%   274.4µ ±  1%  -23.43% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_1[1m])),_range_query_with_1000_steps                   1322.5µ ±  2%   925.7µ ±  1%  -30.00% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_instant_query                                3.075m ±  1%   2.219m ±  1%  -27.83% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_100_steps                  11.788m ±  1%   7.754m ±  2%  -34.22% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_100[1m])),_range_query_with_1000_steps                  85.07m ±  2%   52.19m ±  2%  -38.65% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_instant_query                               49.76m ±  2%   35.74m ±  2%  -28.18% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_100_steps                  214.0m ±  1%   137.1m ±  3%  -35.92% (p=0.002 n=6)
Query/sum_by_(le)(rate(h_2000[1m])),_range_query_with_1000_steps                  1.812 ±  1%    1.000 ±  2%  -44.81% (p=0.002 n=6)
geomean                                                                          10.33m         7.892m        -23.59%

                                                                       │      Prometheus      │               streaming               │
                                                                       │         B/op         │      B/op       vs base               │
Query/a_1,_instant_query                                                         20.68Ki ± 0%    17.75Ki ±  0%  -14.16% (p=0.002 n=6)
Query/a_1,_range_query_with_100_steps                                            21.60Ki ± 0%    18.32Ki ±  0%  -15.19% (p=0.002 n=6)
Query/a_1,_range_query_with_1000_steps                                           26.19Ki ± 0%    22.88Ki ±  0%  -12.65% (p=0.002 n=6)
Query/a_100,_instant_query                                                       169.6Ki ± 0%    151.8Ki ±  0%  -10.53% (p=0.002 n=6)
Query/a_100,_range_query_with_100_steps                                          222.4Ki ± 0%    208.6Ki ±  0%   -6.20% (p=0.002 n=6)
Query/a_100,_range_query_with_1000_steps                                         658.3Ki ± 0%    645.4Ki ±  0%   -1.96% (p=0.002 n=6)
Query/a_2000,_instant_query                                                      3.019Mi ± 0%    2.690Mi ±  0%  -10.89% (p=0.002 n=6)
Query/a_2000,_range_query_with_100_steps                                         3.965Mi ± 0%    3.771Mi ±  1%   -4.90% (p=0.002 n=6)
Query/a_2000,_range_query_with_1000_steps                                        12.34Mi ± 0%    12.14Mi ±  0%   -1.59% (p=0.002 n=6)
Query/nh_1,_instant_query                                                        37.10Ki ± 0%    34.09Ki ±  0%   -8.10% (p=0.002 n=6)
Query/nh_1,_range_query_with_100_steps                                           88.56Ki ± 0%    85.46Ki ±  0%   -3.50% (p=0.002 n=6)
Query/nh_1,_range_query_with_1000_steps                                          520.0Ki ± 0%    517.4Ki ±  0%   -0.48% (p=0.002 n=6)
Query/nh_100,_instant_query                                                      940.2Ki ± 0%    922.1Ki ±  0%   -1.93% (p=0.002 n=6)
Query/nh_100,_range_query_with_100_steps                                         5.568Mi ± 0%    5.563Mi ±  0%   -0.10% (p=0.002 n=6)
Query/nh_100,_range_query_with_1000_steps                                        47.51Mi ± 0%    47.52Mi ±  0%        ~ (p=0.818 n=6)
Query/nh_2000,_instant_query                                                     17.76Mi ± 0%    17.46Mi ±  0%   -1.68% (p=0.002 n=6)
Query/nh_2000,_range_query_with_100_steps                                        111.6Mi ± 0%    111.5Mi ±  0%        ~ (p=0.240 n=6)
Query/nh_2000,_range_query_with_1000_steps                                       977.2Mi ± 0%    976.8Mi ±  0%   -0.05% (p=0.004 n=6)
Query/a_1[1m],_instant_query                                                     21.19Ki ± 0%    17.38Ki ±  0%  -17.98% (p=0.002 n=6)
Query/a_100[1m],_instant_query                                                   132.7Ki ± 0%    109.2Ki ±  0%  -17.69% (p=0.002 n=6)
Query/a_2000[1m],_instant_query                                                  2.291Mi ± 0%    1.903Mi ±  0%  -16.92% (p=0.002 n=6)
Query/rate(a_1[1m]),_instant_query                                               22.47Ki ± 0%    17.82Ki ±  0%  -20.70% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_100_steps                                  23.40Ki ± 1%    18.25Ki ±  0%  -21.98% (p=0.002 n=6)
Query/rate(a_1[1m]),_range_query_with_1000_steps                                 28.21Ki ± 1%    22.90Ki ±  0%  -18.82% (p=0.002 n=6)
Query/rate(a_100[1m]),_instant_query                                             151.8Ki ± 0%    111.5Ki ±  0%  -26.51% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_100_steps                                192.2Ki ± 0%    155.6Ki ±  0%  -19.05% (p=0.002 n=6)
Query/rate(a_100[1m]),_range_query_with_1000_steps                               651.5Ki ± 0%    613.3Ki ±  0%   -5.85% (p=0.002 n=6)
Query/rate(a_2000[1m]),_instant_query                                            2.616Mi ± 0%    1.878Mi ±  0%  -28.21% (p=0.002 n=6)
Query/rate(a_2000[1m]),_range_query_with_100_steps                               3.381Mi ± 0%    2.794Mi ±  1%  -17.36% (p=0.002 n=6)
…